requirements.txt
absl-py==1.0.0
aioice==0.7.5
aiortc==1.2.1
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
av==8.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
black==21.12b0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.2.0
entrypoints==0.3
executing==0.8.2
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.26
google-crc32c==1.3.0
idna==3.3
imageio==2.14.0
imageio-ffmpeg==0.4.5
importlib-resources==5.4.0
imutils==0.5.4
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
moviepy==1.0.3
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.6.3
notebook==6.4.7
numpy @ file:///C:/ci/numpy_and_numpy_base_1634106873763/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
proglog==0.1.9
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
pure-eval==0.2.1
pyarrow==6.0.1
pycparser==2.21
pydeck==0.7.1
pyee==9.0.3
Pygments==2.11.2
pylibsrtp==0.7.0
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
stack-data==0.1.4
streamlit==1.4.0
streamlit-webrtc==0.34.2
terminado==0.12.1
testpath==0.5.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wincertstore==0.2
zipp==3.7.0
